add operationoutcome referenced resources to contained
